+++ CHANGELOG.md
@@ -23,10 +23,14 @@
 8.2
 -----
 *   Bug Fixes
+    *   Fix notifications choose podcasts obscuring the mini player
+        ([#4861](https://github.com/Automattic/pocket-casts-android/pull/4861))
     *   Fix Android 10 full screen player appearing under the status and navigation bars
         ([#4804](https://github.com/Automattic/pocket-casts-android/pull/4804))
     *   Fix the mini player Up Next count overflow
         ([#4826](https://github.com/Automattic/pocket-casts-android/pull/4826))
+    *   Improve help & feedback page padding
+        ([#4858](https://github.com/Automattic/pocket-casts-android/pull/4858)) 
     *   Fix 3-button navigation obstructing the sharing buttons
         ([#4810](https://github.com/Automattic/pocket-casts-android/pull/4810))